refactor transfer, fixes
