+++ xrpl-contract-stdlib/src/submit/mod.rs
@@ -1,1 +1,2 @@
+pub mod amount;
 pub mod inner_objects;
